Fix logic
