+++ projects/kea/fuzz_dhcp_pkt_process4.cc
@@ -7,9 +7,9 @@
 #include <config.h>
 #include <fuzzer/FuzzedDataProvider.h>
 
-#include <dhcp/pkt6.h>
+#include <dhcp/pkt4.h>
 #include <dhcp/libdhcp++.h>
-#include <dhcp6/ctrl_dhcp6_srv.h>
+#include <dhcp4/ctrl_dhcp4_srv.h>
 #include <log/logger_support.h>
 #include <util/filesystem.h>
 
@@ -31,13 +31,17 @@ using namespace isc::util;
 
 namespace isc {
     namespace dhcp {
-        class MyDhcpv6Srv : public ControlledDhcpv6Srv {
+        class MyDhcpv4Srv : public ControlledDhcpv4Srv {
             public:
-                void fuzz_sanityCheck(const Pkt6Ptr& query) {
-                    sanityCheck(query);
+                bool fuzz_accept(const Pkt4Ptr& pkt) {
+                    return accept(pkt);
                 }
 
-                void fuzz_classifyPacket(const Pkt6Ptr& pkt) {
+                static void fuzz_sanityCheck(const Pkt4Ptr& query) {
+                    ControlledDhcpv4Srv::sanityCheck(query);
+                }
+
+                void fuzz_classifyPacket(const Pkt4Ptr& pkt) {
                     classifyPacket(pkt);
                 }
 
@@ -75,21 +79,26 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
         return 0;
     }
 
-    Pkt6Ptr pkt;
-    std::unique_ptr<MyDhcpv6Srv> srv;
+    Pkt4Ptr pkt;
+    std::unique_ptr<MyDhcpv4Srv> srv;
 
     // Package parsing
     try {
-        pkt = Pkt6Ptr(new Pkt6(data, size));
+        pkt = Pkt4Ptr(new Pkt4(data, size));
         pkt->unpack();
     } catch (...) {
         // Early exit if package parsing failed.
         return 0;
     }
 
+    // Configure random value in packet
+    FuzzedDataProvider fdp(data, size);
+    uint8_t typeChoice = fdp.ConsumeIntegralInRange<uint8_t>(0, 8);
+    pkt->setType(static_cast<DHCPMessageType>(typeChoice));
+
     // Server initialisation
     try {
-        srv.reset(new MyDhcpv6Srv());
+        srv.reset(new MyDhcpv4Srv());
         srv->init(path);
     } catch (...) {
         // Early exit if server initialisation failed.
@@ -108,9 +117,23 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
         // Slient exceptions
     }
 
+    // Call accept for packet checking
+    try {
+        srv->fuzz_accept(pkt);
+    } catch (const isc::Exception& e) {
+        // Slient exceptions
+    }
+
     // Call sanityCheck for packet checking
     try {
-        srv->fuzz_sanityCheck(pkt);
+        MyDhcpv4Srv::fuzz_sanityCheck(pkt);
+    } catch (const isc::Exception& e) {
+        // Slient exceptions
+    }
+
+    // Call process functions after the accept and check
+    try {
+        srv->processDhcp4Query(pkt, fdp.ConsumeBool());
     } catch (const isc::Exception& e) {
         // Slient exceptions
     }
